refactor(cache): move Cache class to Plugins namespace
