+++ tests/Unit/Services/InMemorySpellcheckerTest.php
@@ -2,8 +2,8 @@
 
 declare(strict_types=1);
 
-use Peck\Cache;
 use Peck\Config;
+use Peck\Plugins\Cache;
 use Peck\Services\Spellcheckers\InMemorySpellchecker;
 use PhpSpellcheck\Spellchecker\Aspell;
